Tweak font size and fix font scaling in Safari
Safari does not resize the text as the viewport is resized. This hack
gets around that by forcing a repaint on every frame. This does not seem
to impact the browsers CPU usage at all.

One interesting issue is that Safari snaps the font sizes to the
nearest pixel size, so the font sizes tend to jump around a little,
rather than scaling smoothly like Firefox/Chrome.